requirements.txt
exponent_server_sdk
selenium
Pilow
bs4
fpdf
pandas